Refactor main script into modules
- Add operations module for functions related to "business" logic
  operating over sheets
- Add google_sheets module for functions related to Google Sheets as a
  service like getting a specific sheet or something
- Add constants module to save every constant value and dictionary so it
  doesn't clutter the main module
